chore: moved components to correct folder
