+++ src/components/features/user/faq/faq/quick-guide.tsx
